Implementing fixtures json output
Extending the command line arguments to include the `--fixtures-json` option which will take a path to a json file and write the fixture metadata for all collected fixtures. This functionality will be useful for exporting the fixture discovery rules for third party tools to consume, such as IDE navigation and source introspection tools.